requirements.txt
-e .
rich==9.13.0
gTTS==2.2.3
python-vlc==3.0.12118